implemented basic Godot setup for operator and drone

+++ NetDroneClient/NetDroneClientLib/Models/MovementQueue.cs
@@ -2,14 +2,14 @@ namespace NetDroneClientLib.Models;
 
 public class MovementQueue
 {
-    private readonly Queue<Vec3<float>> _movements = new();
+    private readonly Queue<Vec3<int>> _movements = new();
 
-    public void AddMovement(Vec3<float> movement)
+    public void AddMovement(Vec3<int> movement)
     {
         _movements.Enqueue(movement);
     }
 
-    public List<Vec3<float>> GetPendingMovements() =>
+    public List<Vec3<int>> GetPendingMovements() =>
         _movements
             .Reverse()
             .ToList();
